make PlaneSlice API accept u32 instead of usize

+++ src/plane.rs
@@ -827,7 +827,7 @@ impl<'a, T: Pixel> PlaneSlice<'a, T> {
         }
     }
 
-    pub fn subslice(&self, xo: usize, yo: usize) -> PlaneSlice<'a, T> {
+    pub fn subslice(&self, xo: u32, yo: u32) -> PlaneSlice<'a, T> {
         PlaneSlice {
             plane: self.plane,
             x: self.x + xo as isize,
@@ -844,7 +844,7 @@ impl<'a, T: Pixel> PlaneSlice<'a, T> {
     }
 
     /// A slice starting i pixels above the current one.
-    pub fn go_up(&self, i: usize) -> PlaneSlice<'a, T> {
+    pub fn go_up(&self, i: u32) -> PlaneSlice<'a, T> {
         PlaneSlice {
             plane: self.plane,
             x: self.x,
@@ -853,7 +853,7 @@ impl<'a, T: Pixel> PlaneSlice<'a, T> {
     }
 
     /// A slice starting i pixels to the left of the current one.
-    pub fn go_left(&self, i: usize) -> PlaneSlice<'a, T> {
+    pub fn go_left(&self, i: u32) -> PlaneSlice<'a, T> {
         PlaneSlice {
             plane: self.plane,
             x: self.x - i as isize,
@@ -861,7 +861,7 @@ impl<'a, T: Pixel> PlaneSlice<'a, T> {
         }
     }
 
-    pub fn p(&self, add_x: usize, add_y: usize) -> T {
+    pub fn p(&self, add_x: u32, add_y: u32) -> T {
         let new_y = (self.y + add_y as isize + self.plane.cfg.yorigin as isize) as usize;
         let new_x = (self.x + add_x as isize + self.plane.cfg.xorigin as isize) as usize;
         self.plane.data[new_y * self.plane.cfg.stride as usize + new_x]
@@ -869,7 +869,7 @@ impl<'a, T: Pixel> PlaneSlice<'a, T> {
 
     /// Checks if `add_y` and `add_x` lies in the allocated bounds of the
     /// underlying plane.
-    pub fn accessible(&self, add_x: usize, add_y: usize) -> bool {
+    pub fn accessible(&self, add_x: u32, add_y: u32) -> bool {
         let y = (self.y + add_y as isize + self.plane.cfg.yorigin as isize) as u32;
         let x = (self.x + add_x as isize + self.plane.cfg.xorigin as isize) as u32;
         y < self.plane.cfg.alloc_height() && x < self.plane.cfg.stride
@@ -877,14 +877,14 @@ impl<'a, T: Pixel> PlaneSlice<'a, T> {
 
     /// Checks if -`sub_x` and -`sub_y` lies in the allocated bounds of the
     /// underlying plane.
-    pub fn accessible_neg(&self, sub_x: usize, sub_y: usize) -> bool {
+    pub fn accessible_neg(&self, sub_x: u32, sub_y: u32) -> bool {
         let y = self.y - sub_y as isize + self.plane.cfg.yorigin as isize;
         let x = self.x - sub_x as isize + self.plane.cfg.xorigin as isize;
         y >= 0 && x >= 0
     }
 
     /// This version of the function crops off the padding on the right side of the image
-    pub fn row_cropped(&self, y: usize) -> &[T] {
+    pub fn row_cropped(&self, y: u32) -> &[T] {
         let y = (self.y + y as isize + self.plane.cfg.yorigin as isize) as usize;
         let x = (self.x + self.plane.cfg.xorigin as isize) as usize;
         let start = y * self.plane.cfg.stride as usize + x;
@@ -893,7 +893,7 @@ impl<'a, T: Pixel> PlaneSlice<'a, T> {
     }
 
     /// This version of the function includes the padding on the right side of the image
-    pub fn row(&self, y: usize) -> &[T] {
+    pub fn row(&self, y: u32) -> &[T] {
         let y = (self.y + y as isize + self.plane.cfg.yorigin as isize) as usize;
         let x = (self.x + self.plane.cfg.xorigin as isize) as usize;
         let start = y * self.plane.cfg.stride as usize + x;
